cleaned up methods that were used only in tests
